Fixed definition of size_t

+++ src/kernel/netbsd/sys/sysctl.h
@@ -1,5 +1,5 @@
 /* $Id$ */
-/* Copyright (c) 2010 Pierre Pronchery <khorben@defora.org> */
+/* Copyright (c) 2007-2012 Pierre Pronchery <khorben@defora.org> */
 /* This file is part of DeforaOS System libc */
 /* This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -22,7 +22,7 @@
 /* types */
 # ifndef size_t
 #  define size_t size_t
-typedef unsigned int size_t;
+typedef unsigned long size_t;
 # endif
 
 struct clockinfo
